Remove incompatible ashmem.h
Dmitry Antipov noticed there is some inconsistent ioctl definitions
for ashmem in the frameworks/native/include/utils/ashmem.h file.

Specifically, that file defines:

 #define ASHMEM_PIN              _IO(__ASHMEMIOC, 7)
 #define ASHMEM_UNPIN            _IO(__ASHMEMIOC, 8)
 #define ASHMEM_ISPINNED         _IO(__ASHMEMIOC, 9)

Instead of what the kernel, bionic (and apparently everywhere else)
uses:
 #define ASHMEM_PIN     _IOW(__ASHMEMIOC, 7, struct ashmem_pin)
 #define ASHMEM_UNPIN   _IOW(__ASHMEMIOC, 8, struct ashmem_pin)
 #define ASHMEM_GET_PIN_STATUS	_IO(__ASHMEMIOC, 9)

The problematic file doesn't seem to have much git history,
and seems to be a historic leftover. So after checking with
the Android Kernel Team, it was suggested that removing this
file is the right approach.

Change-Id: I6da9989856b74deed2e1bd28d3d7f9f9c70b85a0
Signed-off-by: John Stultz <john.stultz@linaro.org>